modularized a lil bit
