+++ client/components/trainterLibraries/Card.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Card() {
+  return <div>Card</div>;
+}
