+++ client/components/trainterLibraries/Container.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Container() {
+  return <div>Container</div>;
+}
