+++ client/pages/trainer/exercise/[id].tsx
@@ -1,14 +1,14 @@
+import { ExerciesResType } from "@/types/components/libraries";
 import axios from "axios";
 import Image from "next/image";
 import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
-import { ResType } from "../exercises-library";
 
 export default function ExerciseDetails() {
   const router = useRouter();
   const { id } = router.query;
 
-  const [exerciseData, setExerciseData] = useState<ResType>();
+  const [exerciseData, setExerciseData] = useState<ExerciesResType>();
 
   const goBackHandler = () => {
     router.replace("/trainer/exercises-library");
