+++ client/pages/trainer/food-library.tsx
@@ -1,31 +1,9 @@
+import NavigationBtns from "@/components/trainterLibraries/NavigationBtns";
+import { FoodResType } from "@/types/components/libraries";
 import axios from "axios";
 import Link from "next/link";
 import React, { useEffect, useState } from "react";
 import { FaHamburger } from "react-icons/fa";
-
-export type FoodResType = {
-  id: number;
-  dataType: string;
-  description: string;
-  proteinAmount: number;
-  proteinUnit: string;
-  carbohydrateAmount: number;
-  carbohydrateUnit: string;
-  fatTransAmount: number;
-  fatTransUnit: string;
-  fatSaturatedAmount: number;
-  fatSaturatedUnit: string;
-  fatTotalAmount: number;
-  fatTotalUnit: string;
-  sugarsAmount: number;
-  sugarsdUnit: string;
-  sodiumAmount: number;
-  sodiumUnit: string;
-  cholesterolAmount: number;
-  cholesterolUnit: string;
-  energyAmount: number;
-  energylUnit: string;
-};
 
 export default function FoodLibrary() {
   const [foods, setFoods] = useState<FoodResType[]>([]);
@@ -56,23 +34,13 @@ export default function FoodLibrary() {
 
   return (
     <div>
-      <div className="flex justify-between items-center w-full pt-[115px] px-36">
-        <button
-          onClick={prevPage}
-          disabled={currentPage === 1}
-          className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-        >
-          Anterior
-        </button>
-        <div className="text-xl font-bold">{currentPage}</div>
-        <button
-          onClick={nextPage}
-          disabled={foods.length === 0}
-          className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-        >
-          Siguiente
-        </button>
-      </div>
+      <NavigationBtns
+        {...{ currentPage }}
+        {...{ prevPage }}
+        {...{ nextPage }}
+        length={foods.length}
+      />
+
       <div className="flex flex-wrap gap-20">
         {foods.slice(0, 30).map((food) => (
           <Link
@@ -85,24 +53,14 @@ export default function FoodLibrary() {
             <FaHamburger />
           </Link>
         ))}
-        <div className="flex justify-between items-center w-full mt-10 px-36">
-          <button
-            onClick={prevPage}
-            disabled={currentPage === 1}
-            className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-          >
-            Anterior
-          </button>
-          <div className="text-xl font-bold">{currentPage}</div>
-          <button
-            onClick={nextPage}
-            disabled={foods.length === 0}
-            className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-          >
-            Siguiente
-          </button>
-        </div>
       </div>
+
+      <NavigationBtns
+        {...{ currentPage }}
+        {...{ prevPage }}
+        {...{ nextPage }}
+        length={foods.length}
+      />
     </div>
   );
 }
